Add image files
These designs were created using HTML, CSS, and JS. 
I tried to maintain consistency as best as I could, but some edits may still be necessary. 
The designs are based on the Figma design that we submitted for 4a. 
I would like to discuss how we can integrate them with the backend.